feat(features): Require polls as given
"Requérir la présence à priori de scrutins."

Another brick in setting up the contexts
we'll want for our scenarios.
Language detection is a dummy for now.
But we're add a rather fault-intolerant but working YAML perser
to use with step pystrings.
I18N support is overall made of tape (not even cardboard).