conf(features): Tweak behave some more.
Make sure when using print() in step defs to end it with a `\n`,
This is a pitfall of behave, I guess?

We can set the format to plain, using the --format plain CLI option,
since the `format = plain` in the `.behaverc` changes nothing.
When in plain format, there are no colors and our print() appears.
We tried behave 1.2.6 and 1.2.7-dev2.
This may well be solved by some tweaking.
Found out about `aloe`, and `pytest-bdd`.
This may well be solved by some experimenting.